VolatileDB: split off state-related functions in Impl.State
