+++ ouroboros-consensus/ouroboros-consensus.cabal
@@ -152,6 +152,7 @@ library
                        Ouroboros.Consensus.Storage.VolatileDB.API
                        Ouroboros.Consensus.Storage.VolatileDB.FileInfo
                        Ouroboros.Consensus.Storage.VolatileDB.Impl
+                       Ouroboros.Consensus.Storage.VolatileDB.Impl.State
                        Ouroboros.Consensus.Storage.VolatileDB.Index
                        Ouroboros.Consensus.Storage.VolatileDB.Types
                        Ouroboros.Consensus.Storage.VolatileDB.Util
